requirements.txt
Django
djangorestframework
gunicorn
python-dotenv
requests